fix: extract delivery status tests to keep tests.rs under 1000 lines (#62)
