+++ crates/rs-core/src/db/mod.rs
@@ -21,6 +21,9 @@ mod template_tests;
 #[cfg(test)]
 mod delivery_log_tests;
 
+#[cfg(test)]
+mod delivery_status_tests;
+
 /// Create a SQLite connection pool.
 pub async fn create_pool(db_path: &Path) -> Result<SqlitePool> {
     let url = format!("sqlite:{}?mode=rwc", db_path.display());
